[#14] PostLockService 삭제
